Define separate chain interface for pkg/coordinator
The functions used by the coordinator package should be extracted from
the pkg/tbtc to make tha packages implementation clean, as only the
functions that are used in the given package should be defined in the
chain interface for the package.

Refs: https://github.com/keep-network/keep-core/issues/3632
