+++ pkg/tbtc/chain.go
@@ -194,22 +194,6 @@ var (
 // BridgeChain defines the subset of the TBTC chain interface that pertains
 // specifically to the tBTC Bridge operations.
 type BridgeChain interface {
-	// GetDepositRequest gets the on-chain deposit request for the given
-	// funding transaction hash and output index. Returns an error if the
-	// deposit was not found.
-	GetDepositRequest(
-		fundingTxHash bitcoin.Hash,
-		fundingOutputIndex uint32,
-	) (*DepositChainRequest, error)
-
-	// PastNewWalletRegisteredEvents fetches past new wallet registered events
-	// according to the provided filter or unfiltered if the filter is nil. Returned
-	// events are sorted by the block number in the ascending order, i.e. the
-	// latest event is at the end of the slice.
-	PastNewWalletRegisteredEvents(
-		filter *NewWalletRegisteredEventFilter,
-	) ([]*NewWalletRegisteredEvent, error)
-
 	// GetWallet gets the on-chain data for the given wallet. Returns an error
 	// if the wallet was not found.
 	GetWallet(walletPublicKeyHash [20]byte) (*WalletChainData, error)
@@ -217,20 +201,6 @@ type BridgeChain interface {
 	// ComputeMainUtxoHash computes the hash of the provided main UTXO
 	// according to the on-chain Bridge rules.
 	ComputeMainUtxoHash(mainUtxo *bitcoin.UnspentTransactionOutput) [32]byte
-
-	// BuildDepositKey calculates a deposit key for the given funding transaction
-	// which is a unique identifier for a deposit on-chain.
-	BuildDepositKey(fundingTxHash bitcoin.Hash, fundingOutputIndex uint32) *big.Int
-
-	// GetDepositParameters gets the current value of parameters relevant
-	// for the depositing process.
-	GetDepositParameters() (
-		dustThreshold uint64,
-		treasuryFeeDivisor uint64,
-		txMaxFee uint64,
-		revealAheadPeriod uint32,
-		err error,
-	)
 
 	// GetPendingRedemptionRequest gets the on-chain pending redemption request
 	// for the given wallet public key hash and redeemer output script.
@@ -304,21 +274,6 @@ type DepositChainRequest struct {
 	SweptAt     time.Time
 }
 
-// NewWalletRegisteredEvent represents a new wallet registered event.
-type NewWalletRegisteredEvent struct {
-	EcdsaWalletID       [32]byte
-	WalletPublicKeyHash [20]byte
-	BlockNumber         uint64
-}
-
-// NewWalletRegisteredEventFilter is a component allowing to filter NewWalletRegisteredEvent.
-type NewWalletRegisteredEventFilter struct {
-	StartBlock          uint64
-	EndBlock            *uint64
-	EcdsaWalletID       [][32]byte
-	WalletPublicKeyHash [][20]byte
-}
-
 // WalletChainData represents wallet data stored on-chain.
 type WalletChainData struct {
 	EcdsaWalletID                          [32]byte
@@ -362,16 +317,6 @@ type WalletCoordinatorChain interface {
 	GetWalletLock(
 		walletPublicKeyHash [20]byte,
 	) (time.Time, WalletActionType, error)
-
-	// SubmitDepositSweepProposalWithReimbursement submits a deposit sweep
-	// proposal to the chain. It reimburses the gas cost to the caller.
-	SubmitDepositSweepProposalWithReimbursement(
-		proposal *DepositSweepProposal,
-	) error
-
-	// GetDepositSweepMaxSize gets the maximum number of deposits that can
-	// be part of a deposit sweep proposal.
-	GetDepositSweepMaxSize() (uint16, error)
 
 	// OnRedemptionProposalSubmitted registers a callback that is invoked when
 	// an on-chain notification of the redemption proposal submission is seen.
